refactor: quest command service 통합
가독성을 위해 update 관련 로직 삭제에 따른 commandService로 로직을 통합했습니다.
