+++ src/main/java/com/groom/orbit/goal/app/command/QuestCommandService.java
@@ -5,9 +5,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.groom.orbit.ai.app.VectorService;
 import com.groom.orbit.ai.app.dto.CreateVectorDto;
+import com.groom.orbit.ai.app.dto.UpdateVectorQuestDto;
 import com.groom.orbit.common.dto.CommonSuccessDto;
 import com.groom.orbit.goal.app.MemberGoalService;
 import com.groom.orbit.goal.app.dto.request.CreateQuestRequestDto;
+import com.groom.orbit.goal.app.dto.request.UpdateQuestRequestDto;
 import com.groom.orbit.goal.app.dto.response.CreateQuestResponse;
 import com.groom.orbit.goal.app.query.QuestQueryService;
 import com.groom.orbit.goal.dao.QuestRepository;
@@ -48,9 +50,29 @@ public class QuestCommandService {
     return CommonSuccessDto.fromEntity(true);
   }
 
+  public CommonSuccessDto updateQuest(Long memberId, Long questId, UpdateQuestRequestDto dto) {
+    Quest quest = questQueryService.findQuest(questId);
+    quest.validateMember(memberId);
+
+    updateVector(memberId, dto, quest);
+    quest.update(dto.title(), dto.isComplete(), dto.deadline());
+
+    return new CommonSuccessDto(true);
+  }
+
   private void saveVector(Long memberId, CreateQuestRequestDto dto) {
     CreateVectorDto vectorDto =
         CreateVectorDto.builder().memberId(memberId).quest(dto.title()).build();
     vectorService.save(vectorDto);
+  }
+
+  private void updateVector(Long memberId, UpdateQuestRequestDto dto, Quest quest) {
+    UpdateVectorQuestDto updateDto =
+        UpdateVectorQuestDto.builder()
+            .memberId(memberId)
+            .quest(quest.getTitle())
+            .newQuest(dto.title())
+            .build();
+    vectorService.updateQuest(updateDto);
   }
 }
